reorganize documentation
Take advantage of new functionality in Racket tools to generate a
Rhombus-oriented view of all documentation.

For now, some information is duplicated in "info.rkt" `scribblings`
entries and in `title` category arguments. When we no longer care how
nicely Rhombus documentation is organized in Racket v9.0 and earlier,
we can shift to a single copy at `title`.